Convert variable from int to float, in order to accept floating value as well

+++ 11_first-try-except.py
@@ -1,6 +1,7 @@
 number = raw_input('Enter a number to find a square : ')
 try :
-  actualNumber = int(number)**2
-  print 'Square of the number is', actualNumber
+    # In order to accept floating numbers, we are converting the varibale to float.
+    actualNumber = float(number)**2
+    print 'Square of the number is', actualNumber
 except :
   print 'Instead of typing number you entered -', number
